fix(uplc): clear the remaining cppcheck 2.7 MISRA findings
Cast both operands to int in the parser's char comparisons (10.4), drop the redundant initializer on two fully-assigned local arrays (9.3), and widen the constr-tag literals before the addition rather than the composite result (10.7). Verified against cppcheck 2.7, the version CI runs.

+++ lib/src/uplc/data/uplc_data.c
@@ -843,7 +843,7 @@ alternative_to_tag(uint64_t alternative)
 
   if (alternative <= 127U)
   {
-    return (1280U - 7U) + alternative;
+    return ((uint64_t)1280U - (uint64_t)7U) + alternative;
   }
 
   return 102U;
